refactor: move NetworkMonitor and initialize Sync
- Moved `NetworkMonitor` from `core.data.utils` to `core.domain.utils`.
- Initialized `Sync` in `POSApp`'s `onCreate` method.
