+++ app/src/main/java/com/casecode/pos/ui/MainAppState.kt
@@ -28,7 +28,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.casecode.pos.MainAuthUiState
-import com.casecode.pos.core.data.utils.NetworkMonitor
+import com.casecode.pos.core.domain.utils.NetworkMonitor
 import com.casecode.pos.feature.inventory.navigation.InventoryRoute
 import com.casecode.pos.feature.profile.navigateToProfile
 import com.casecode.pos.feature.purchase.navigation.PurchaseRoute
